# openapi.yaml
openapi: 3.0.0
info:
  title: Riot Stats API
  version: 1.0.0
  description: >
    API REST pour trois collections MongoDB: match_ids, players, match_data.
    Les identifiants sont:
    - match_ids: matchId
    - players: puuid
    - match_data: metadata.matchId
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/match-data:
    get:
      tags: [match-data]
      summary: Lister les match_data
      operationId: listMatchData
      description: >
        Liste paginée, filtrable par matchId, platformId, queueId, gameVersion et
        bornes de timestamps/durée.
      parameters:
        - in: query
          name: matchId
          schema: { type: string }
        - in: query
          name: platformId
          schema: { type: string, example: "NA1" }
        - in: query
          name: queueId
          schema: { type: integer, example: 440 }
        - in: query
          name: gameVersion
          schema: { type: string, example: "15.16.704.6097" }
        - in: query
          name: startTimeFrom
          schema: { type: integer, format: int64, description: "epoch ms" }
        - in: query
          name: startTimeTo
          schema: { type: integer, format: int64 }
        - in: query
          name: durationMin
          schema: { type: integer, description: "seconds" }
        - in: query
          name: durationMax
          schema: { type: integer, description: "seconds" }
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: sort
          schema: { type: string, example: "info.gameEndTimestamp,desc" }
      responses:
        "200":
          description: Page de match_data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageMatchData" }
    post:
      tags: [match-data]
      summary: Créer un document match_data
      operationId: createMatchData
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MatchData" }
      responses:
        "201":
          description: Créé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchData" }
        "409":
          description: Conflit (déjà présent pour ce matchId)
  /api/match-data/bulk:
    post:
      tags: [match-data]
      summary: Upsert en lot (clé = metadata.matchId)
      operationId: bulkUpsertMatchData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/MatchData" }
      responses:
        "200":
          description: Résultat du bulk
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted: { type: integer }
                  updated: { type: integer }
                  total:   { type: integer }
  /api/match-data/{matchId}:
    get:
      tags: [match-data]
      summary: Récupérer par matchId (clé métier)
      operationId: getMatchDataByMatchId
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: string, example: "NA1_5348438296" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchData" }
          links:
            playersInMatch:
              operationId: getPlayersForMatch
              parameters:
                matchId: "$response.body#/metadata/matchId"
        "404":
          description: Introuvable
    delete:
      tags: [match-data]
      summary: Supprimer par matchId
      operationId: deleteMatchDataByMatchId
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Supprimé
        "404":
          description: Introuvable

  # Relation: match_data -> players list
  /api/match-data/{matchId}/players:
    get:
      tags: [match-data, players]
      summary: Lister les players impliqués dans ce match (documents players)
      operationId: getPlayersForMatch
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Player" }
        "404":
          description: Introuvable
  /api/match-data/participants/by-puuid/{puuid}:
    get:
      tags: [match-data]
      summary: Chercher toutes les parties d’un joueur (PUUID)
      operationId: getMatchesByPuuid
      parameters:
        - in: path
          name: puuid
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Liste des parties où ce PUUID apparaît
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageMatchData" }
          links:
            playerDoc:
              operationId: getPlayerByPuuid
              parameters:
                puuid: "$request.path.puuid"
  /api/match-data/stats/durations:
    get:
      tags: [match-data]
      summary: Stats de durée des parties
      operationId: matchDurationsStats
      parameters:
        - in: query
          name: queueId
          schema: { type: integer }
        - in: query
          name: platformId
          schema: { type: string }
        - in: query
          name: startTimeFrom
          schema: { type: integer, format: int64 }
        - in: query
          name: startTimeTo
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Min/Max/Avg
          content:
            application/json:
              schema:
                type: object
                properties:
                  min: { type: integer, example: 1023 }
                  max: { type: integer, example: 3123 }
                  avg: { type: number, format: float, example: 1682.4 }
  /api/match-data/stats/champions:
    get:
      tags: [match-data]
      summary: Fréquences de champions (top N)
      operationId: championFrequency
      parameters:
        - in: query
          name: byTeam
          schema: { type: boolean, default: false }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        "200":
          description: Top champions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    championId: { type: integer }
                    count: { type: integer }
  /api/match-data/stats/winrate-by-champion:
    get:
      tags: [match-data]
      summary: Winrate par champion (optionnellement filtré)
      operationId: winrateByChampion
      parameters:
        - in: query
          name: queueId
          schema: { type: integer }
        - in: query
          name: platformId
          schema: { type: string }
      responses:
        "200":
          description: Winrate par champion
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    championId: { type: integer }
                    games: { type: integer }
                    wins: { type: integer }
                    winrate: { type: number, format: float, example: 52.3 }
components:
  schemas:
    Tier:
      type: string
      description: Palier de league
      enum: [IRON, BRONZE, SILVER, GOLD, PLATINUM, EMERALD, DIAMOND, MASTER, GRANDMASTER, CHALLENGER]
      example: PLATINUM
    Rank:
      type: string
      description: Division au sein du palier
      enum: [I, II, III, IV]
      example: I

    ##########################
    # MATCH IDS SCHEMAS
    ##########################
...
    Player:
      type: object
      additionalProperties: false
      properties:
        puuid:
          type: string
          description: Riot PUUID (identifiant unique global)
          example: "CEAj3J8ZSyHIHaARmGnWDbQJ34-QMCgrdueWLRKvPUSEaSAtvkKBMRM1QP5FojYGOnNb6Gx-6eqYMQ"
        tier: { $ref: "#/components/schemas/Tier" }
        rank: { $ref: "#/components/schemas/Rank" }
        leaguePoints: { type: integer, minimum: 0, example: 72 }
        wins: { type: integer, minimum: 0, example: 106 }
        losses: { type: integer, minimum: 0, example: 109 }
        veteran: { type: boolean, example: false }
        inactive: { type: boolean, example: false }
        freshBlood: { type: boolean, example: false }
      required: [puuid, tier, rank, leaguePoints, wins, losses, veteran, inactive, freshBlood]
...
    MatchDataMetadata:
      type: object
      additionalProperties: false
      properties:
        dataVersion: { type: string, example: "2" }
        matchId: { type: string, example: "NA1_5348438296" }
        participants:
          type: array
          items: { type: string, description: "PUUID" }
      required: [dataVersion, matchId, participants]
    MatchDataTeamObjectives:
      type: object
      additionalProperties: true
      properties:
        first: { type: boolean }
        kills: { type: integer }
    MatchDataTeamFeats:
      type: object
      additionalProperties: true
      description: "Riot feats map, ex: FIRST_BLOOD, FIRST_TURRET…"
      properties:
        FIRST_BLOOD:
          type: object
          properties:
            featState: { type: integer }
    MatchDataTeam:
      type: object
      additionalProperties: false
      properties:
        teamId: { type: integer, enum: [100, 200] }
        win: { type: boolean }
        bans:
          type: array
          items:
            type: object
            properties:
              championId: { type: integer }
              pickTurn: { type: integer }
        feats:
          $ref: "#/components/schemas/MatchDataTeamFeats"
        objectives:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/MatchDataTeamObjectives"
    MatchDataParticipant:
      type: object
      description: "Participant unique d’une partie"
      additionalProperties: true
      properties:
        puuid: { type: string }
        participantId: { type: integer, minimum: 1, maximum: 10 }
        teamId: { type: integer, enum: [100, 200] }
        championId: { type: integer }
        championName: { type: string }
        individualPosition: { type: string, example: "TOP" }
        teamPosition: { type: string, example: "TOP" }
        kills: { type: integer }
        deaths: { type: integer }
        assists: { type: integer }
        win: { type: boolean }
        goldEarned: { type: integer }
        totalDamageDealtToChampions: { type: integer }
        totalMinionsKilled: { type: integer }
        summonerName: { type: string }
        summonerId: { type: string }
        riotIdGameName: { type: string }
        riotIdTagline: { type: string }
        itemsPurchased: { type: integer }
        item0: { type: integer }
        item1: { type: integer }
        item2: { type: integer }
        item3: { type: integer }
        item4: { type: integer }
        item5: { type: integer }
        item6: { type: integer }
        perks:
          type: object
          additionalProperties: true
    MatchDataInfo:
      type: object
      additionalProperties: false
      properties:
        endOfGameResult: { type: string, example: "GameComplete" }
        gameCreation: { type: integer, format: int64, description: "epoch ms" }
        gameDuration: { type: integer, description: "seconds" }
        gameEndTimestamp: { type: integer, format: int64, description: "epoch ms" }
        gameId: { type: integer, format: int64 }
        gameMode: { type: string, example: "CLASSIC" }
        gameName: { type: string }
        gameStartTimestamp: { type: integer, format: int64, description: "epoch ms" }
        gameType: { type: string, example: "MATCHED_GAME" }
        gameVersion: { type: string, example: "15.16.704.6097" }
        mapId: { type: integer, example: 11 }
        platformId: { type: string, example: "NA1" }
        queueId: { type: integer, example: 440, description: "e.g., Flex 5v5" }
        participants:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/MatchDataParticipant" }
        teams:
          type: array
          items: { $ref: "#/components/schemas/MatchDataTeam" }
        tournamentCode: { type: string }
    MatchData:
      type: object
      additionalProperties: false
      properties:
        metadata: { $ref: "#/components/schemas/MatchDataMetadata" }
        info: { $ref: "#/components/schemas/MatchDataInfo" }
      required: [metadata, info]
    PageMatchData:
      type: object
      properties:
        content:
          type: array
          items: { $ref: "#/components/schemas/MatchData" }
        page: { type: integer, example: 0 }
        size: { type: integer, example: 50 }
        totalElements: { type: integer, example: 1234 }
        totalPages: { type: integer, example: 25 }
        sort: { type: string, example: "info.gameEndTimestamp,desc" }